App authentication for GitHub API
GitHub recently reduced rate limits for unauthenticated calls to
their API. If an app needs to make unauthenticated calls to the
API but require a high limit(which is almost always the case since
now the default limit is 60 calls per hour), the app needs to send
its client_id and client_secret in the query string. This patch adds
the necessary authentication mechanism and makes it the default
since it makes more sense then the old user credential based auth.